requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==4.9.0
    # via jupyter-server
appnope==0.1.4
    # via
    #   ipykernel
    #   ipython
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==3.0.0
    # via stack-data
async-lru==2.0.5
    # via jupyterlab
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
babel==2.17.0
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.13.4
    # via nbconvert
black==25.1.0
    # via ipython
bleach[css]==6.2.0
    # via nbconvert
certifi==2025.4.26
    # via requests
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via black
comm==0.2.2
    # via ipykernel
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
exceptiongroup==1.2.2
    # via anyio
executing==2.2.0
    # via stack-data
fastjsonschema==2.21.1
    # via nbformat
fonttools==4.57.0
    # via matplotlib
fqdn==1.5.1
    # via jsonschema
idna==3.10
    # via
    #   anyio
    #   jsonschema
    #   requests
importlib-metadata==8.7.0
    # via
    #   jupyter-client
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
importlib-resources==6.5.2
    # via matplotlib
ipykernel==6.29.4
    # via
    #   -r requirements.in
    #   jupyterlab
ipython==8.0.0
    # via
    #   -r requirements.in
    #   ipykernel
isoduration==20.11.0
    # via jsonschema
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
joblib==1.5.0
    # via scikit-learn
json5==0.12.0
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema[format-nongpl]==4.23.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter-client==8.6.3
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.12.0
    # via jupyter-server
jupyter-lsp==2.2.5
    # via jupyterlab
jupyter-server==2.15.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.0.9
    # via
    #   -r requirements.in
    #   notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.27.3
    # via
    #   jupyterlab
    #   notebook
kiwisolver==1.4.7
    # via matplotlib
markupsafe==3.0.2
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.7.1
    # via -r requirements.in
matplotlib-inline==0.1.6
    # via
    #   -r requirements.in
    #   ipykernel
    #   ipython
mistune==3.1.3
    # via nbconvert
mypy-extensions==1.1.0
    # via black
nbclient==0.10.2
    # via nbconvert
nbconvert==7.16.6
    # via
    #   -r requirements.in
    #   jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.0.8
    # via -r requirements.in
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==1.24.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
overrides==7.7.0
    # via jupyter-server
packaging==25.0
    # via
    #   black
    #   ipykernel
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
pandas==2.0.3
    # via -r requirements.in
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==11.2.1
    # via matplotlib
platformdirs==4.3.8
    # via
    #   black
    #   jupyter-core
prometheus-client==0.21.1
    # via jupyter-server
prompt-toolkit==3.0.51
    # via ipython
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pygments==2.19.1
    # via
    #   ipython
    #   nbconvert
pyparsing==3.2.3
    # via matplotlib
pyrtlsdr==0.3.0
    # via -r requirements.in
pyrtlsdrlib==0.0.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
    #   pandas
python-json-logger==3.3.0
    # via jupyter-events
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via jupyter-events
pyzmq==26.4.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.32.3
    # via jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.2.2
    # via -r requirements.in
scipy==1.13.1
    # via
    #   -r requirements.in
    #   scikit-learn
send2trash==1.8.3
    # via jupyter-server
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
threadpoolctl==3.6.0
    # via scikit-learn
tinycss2==1.4.0
    # via bleach
tomli==2.2.1
    # via
    #   black
    #   jupyterlab
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20241206
    # via arrow
typing-extensions==4.13.2
    # via
    #   anyio
    #   async-lru
    #   beautifulsoup4
    #   black
    #   mistune
    #   python-json-logger
    #   referencing
tzdata==2025.2
    # via pandas
uri-template==1.3.0
    # via jsonschema
urllib3==2.4.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==24.11.1
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
zipp==3.21.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
